Cleanup the scenes
